feat: move validation to another package
