+++ pkg/storage/internal/actions/create.go
@@ -23,6 +23,14 @@ type FundingResult struct {
 	Fee            uint64
 }
 
+type CreateActionParams struct {
+}
+
+func FromValidCreateActionArgs(args *wdk.ValidCreateActionArgs) CreateActionParams {
+	// TODO: use only the necessary fields (no redundant fields)
+	return CreateActionParams{}
+}
+
 type Funder interface {
 	Fund(ctx context.Context, tx *sdk.Transaction, userID int) (*FundingResult, error)
 }
@@ -40,64 +48,10 @@ func newCreateAction(logger *slog.Logger, funder Funder) *create {
 	}
 }
 
-func (c *create) Create(auth wdk.AuthID, args wdk.ValidCreateActionArgs) (*wdk.StorageCreateActionResult, error) {
-	if auth.UserID == nil {
-		return nil, fmt.Errorf("missing user ID")
-	}
-	if err := c.validateArgs(&args); err != nil {
-		return nil, err
-	}
-
+func (c *create) Create(auth wdk.AuthID, args CreateActionParams) (*wdk.StorageCreateActionResult, error) {
 	result, err := c.funder.Fund(context.Background(), nil, *auth.UserID)
 	if err != nil {
 		return nil, fmt.Errorf("funding failed: %w", err)
 	}
 	panic(fmt.Errorf("not implemented %v", result))
-}
-
-func (c *create) validateArgs(args *wdk.ValidCreateActionArgs) error {
-	deducedIsSendWith := len(args.Options.SendWith) > 0
-	if args.IsSendWith != deducedIsSendWith {
-		return fmt.Errorf("inconsistent IsSendWith with Options.SendWith")
-	}
-
-	deducedIsRemixChange := !args.IsSendWith && len(args.Inputs) == 0 && len(args.Outputs) == 0
-	if args.IsRemixChange != deducedIsRemixChange {
-		return fmt.Errorf("inconsistent IsRemixChange with IsSendWith and Inputs and Outputs")
-	}
-
-	deducedIsNewTx := args.IsRemixChange || len(args.Inputs) > 0 || len(args.Outputs) > 0
-	if args.IsNewTx != deducedIsNewTx {
-		return fmt.Errorf("inconsistent IsNewTx with IsRemixChange and Inputs and Outputs")
-	}
-
-	if !args.IsNewTx {
-		return fmt.Errorf("create action is meant to create a new transaction")
-	}
-
-	deducedIsSignAction := args.IsNewTx && !args.Options.SignAndProcess && containsNilUnlockingScript(args.Inputs)
-	if args.IsSignAction != deducedIsSignAction {
-		return fmt.Errorf("inconsistent IsSignAction with IsNewTx and Options.SignAndProcess and Inputs.UnlockingScript")
-	}
-
-	deducedIsDelayed := bool(args.Options.AcceptDelayedBroadcast)
-	if args.IsDelayed != deducedIsDelayed {
-		return fmt.Errorf("inconsistent IsDelayed with Options.AcceptDelayedBroadcast")
-	}
-
-	deducedIsNoSend := bool(args.Options.NoSend)
-	if args.IsNoSend != deducedIsNoSend {
-		return fmt.Errorf("inconsistent IsNoSend with Options.NoSend")
-	}
-
-	return nil
-}
-
-func containsNilUnlockingScript(inputs []wdk.ValidCreateActionInput) bool {
-	for _, input := range inputs {
-		if input.UnlockingScript == nil {
-			return true
-		}
-	}
-	return false
 }
