+++ pkg/storage/provider.go
@@ -2,6 +2,7 @@ package storage
 
 import (
 	"fmt"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/validate"
 	"log/slog"
 
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/defs"
@@ -122,7 +123,14 @@ func (p *Provider) FindOrInsertUser(identityKey string) (*wdk.FindOrInsertUserRe
 
 // CreateAction Storage level processing for wallet `createAction`.
 func (p *Provider) CreateAction(auth wdk.AuthID, args wdk.ValidCreateActionArgs) (*wdk.StorageCreateActionResult, error) {
-	res, err := p.actions.Create(auth, args)
+	if auth.UserID == nil {
+		return nil, fmt.Errorf("missing user ID")
+	}
+	if err := validate.ValidCreateActionArgs(&args); err != nil {
+		return nil, err
+	}
+
+	res, err := p.actions.Create(auth, actions.FromValidCreateActionArgs(&args))
 	if err != nil {
 		return nil, fmt.Errorf("failed to create action: %w", err)
 	}
